--- http/src/main/java/io/micronaut/http/originatingips/HttpHeaderRequestIpAddressesResolverConfiguration.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2017-2019 original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.micronaut.http.originatingips;
-
-import io.micronaut.core.util.Toggleable;
-
-import javax.annotation.Nonnull;
-
-/**
- * {@link RequestIpAddressesResolver} implementation which resolves IP Address from an HTTP Header.
- * @author Sergio del Amo
- * @since 1.2.0
- */
-public interface HttpHeaderRequestIpAddressesResolverConfiguration extends Toggleable {
-
-    /**
-     *
-     * @return HTTP Header name.
-     */
-    @Nonnull
-    String getHeaderName();
-
-    /**
-     *
-     * @return The delimiter used eg. {@value ,} to separate IPs in the Http Header value.
-     */
-    @Nonnull
-    String getDelimiter();
-}
