+++ http/src/main/java/io/micronaut/http/util/HttpHeaderRequestIpAddressesResolverConfigurationProperties.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.http.originatingips;
+package io.micronaut.http.util;
 
 import io.micronaut.context.annotation.ConfigurationProperties;
 import io.micronaut.context.annotation.Requires;
