+++ http/src/test/groovy/io/micronaut/http/util/HttpHeaderRequestIpAddressesResolverSpec.groovy
@@ -1,4 +1,4 @@
-package io.micronaut.http.originatingips
+package io.micronaut.http.util
 
 import io.micronaut.context.ApplicationContext
 import io.micronaut.http.HttpHeaders
@@ -7,27 +7,34 @@ import spock.lang.AutoCleanup
 import spock.lang.Shared
 import spock.lang.Specification
 
-class RequestCompositeRequestIpAddressResolverSpec extends Specification {
+class HttpHeaderRequestIpAddressesResolverSpec extends Specification {
 
     @Shared
     @AutoCleanup
     ApplicationContext applicationContext = ApplicationContext.run()
 
-    void "IpAddressesResolverAggregator aggregates every IpAddressesResolver and returns a list of ips"() {
-        given:
-        CompositeRequestIpAddressResolver resolver = applicationContext.getBean(CompositeRequestIpAddressResolver)
+    void "Bean HeaderIpAddressesResolver exists"() {
+        expect:
+        applicationContext.containsBean(HttpHeaderRequestIpAddressesResolver)
+    }
+
+    void "can resolve a list of IPS from X-Forwarded-For header value"() {
+        when:
+        HttpHeaderRequestIpAddressesResolver resolver = applicationContext.getBean(HttpHeaderRequestIpAddressesResolver)
+
+        then:
+        noExceptionThrown()
+
+        when:
         def headers = Stub(HttpHeaders) {
             get(_) >> '34.202.241.227, 10.32.108.32'
         }
         def request = Stub(HttpRequest) {
             getHeaders() >> headers
-            getRemoteAddress() >> new InetSocketAddress(8080)
         }
-
-        when:
         List<String> result = resolver.requestIpAddresses(request)
 
         then:
-        result.sort() == ['34.202.241.227', '10.32.108.32', '0.0.0.0'].sort()
+        result == ['34.202.241.227', '10.32.108.32']
     }
 }
