+++ http/src/test/groovy/io/micronaut/http/util/RemoteAddressRequestIpAddressesResolverSpec.groovy
@@ -1,7 +1,8 @@
-package io.micronaut.http.originatingips
+package io.micronaut.http.util
 
 import io.micronaut.context.ApplicationContext
 import io.micronaut.http.HttpRequest
+import io.micronaut.http.util.RemoteAddressRequestIpAddressesResolver
 import spock.lang.AutoCleanup
 import spock.lang.Shared
 import spock.lang.Specification
