+++ http/src/test/groovy/io/micronaut/http/util/RequestCompositeRequestIpAddressResolverSpec.groovy
@@ -1,4 +1,4 @@
-package io.micronaut.http.originatingips
+package io.micronaut.http.util
 
 import io.micronaut.context.ApplicationContext
 import io.micronaut.http.HttpHeaders
@@ -7,34 +7,27 @@ import spock.lang.AutoCleanup
 import spock.lang.Shared
 import spock.lang.Specification
 
-class HttpHeaderRequestIpAddressesResolverSpec extends Specification {
+class RequestCompositeRequestIpAddressResolverSpec extends Specification {
 
     @Shared
     @AutoCleanup
     ApplicationContext applicationContext = ApplicationContext.run()
 
-    void "Bean HeaderIpAddressesResolver exists"() {
-        expect:
-        applicationContext.containsBean(HttpHeaderRequestIpAddressesResolver)
-    }
-
-    void "can resolve a list of IPS from X-Forwarded-For header value"() {
-        when:
-        HttpHeaderRequestIpAddressesResolver resolver = applicationContext.getBean(HttpHeaderRequestIpAddressesResolver)
-
-        then:
-        noExceptionThrown()
-
-        when:
+    void "IpAddressesResolverAggregator aggregates every IpAddressesResolver and returns a list of ips"() {
+        given:
+        CompositeRequestIpAddressResolver resolver = applicationContext.getBean(CompositeRequestIpAddressResolver)
         def headers = Stub(HttpHeaders) {
             get(_) >> '34.202.241.227, 10.32.108.32'
         }
         def request = Stub(HttpRequest) {
             getHeaders() >> headers
+            getRemoteAddress() >> new InetSocketAddress(8080)
         }
+
+        when:
         List<String> result = resolver.requestIpAddresses(request)
 
         then:
-        result == ['34.202.241.227', '10.32.108.32']
+        result.sort() == ['34.202.241.227', '10.32.108.32', '0.0.0.0'].sort()
     }
 }
